sdm845-common: light: clean up LED blink overrides
The current LED blink override algorithm has never worked as
intended, despite several attempts of fixing it in the past. It
has the following drawbacks, in conjunction with the unified
kernel LED DT bindings:

* the default duty-pcts are way too bright compared to stock
* pause-lo and pause-hi values cannot be overridden properly
  from light HAL
* the overall blink effect makes it almost impossible to
  calculate and rewrite from light HAL to fulfill the on/off
  time period request in lightState. The blink frequency is
  often off, either too fast/slow, or has unnatural transitions

Therefore, clean up the LED blink override code and simply turning
blink on and off to use the stock blink effect from Xiaomi. It has
the following benefits:

* every device will have its own duty-pcts set properly to have
  the most suitable brightness value and ramp steps
* if needed, future override algorithms should take into account
  the difference in idx-len and ramp-step-ms on different devices

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>